19/04/2016 commit

+++ application/controllers/branch_manager_controller.php
@@ -39,7 +39,8 @@ class Branch_manager_controller extends CI_Controller {
 	public function NewRegistration()
 	{
 		$user=$this->session->userdata("userid");
-		if($user!=''){
+		$role_type=$this->session->userdata("role_type");
+		if($user!="" && $role_type=="Branch_Admin"){
 			$data["loginPersonInfo"]=$this->admin_model->getPersonDetails($user);
 			$this->load->view('header',$data);
 			$this->load->view('Registration');
@@ -80,15 +81,27 @@ public function ViewApplication(){
 	{
 			//echo "Delete all record";
 		$application_id=$this->input->post("order");
-		if($application_id){
-		for($i=0;$i<sizeof($application_id);$i++) {
-			# code...
-			//echo $application_id[$i];
-			$this->branch_manager_model->deleteApplication($application_id[$i]);
-		}
-	}
+		$p=$this->input->post("submit");
+		//echo $p;
+		if($p=="delete"){
+		
+			if($application_id){
+					for($i=0;$i<sizeof($application_id);$i++) {
+						$this->branch_manager_model->deleteApplication($application_id[$i]);
+					}
+			}
 
-		redirect('branch_manager_controller/viewApplication');
+		}
+		else{
+				
+			if($application_id){
+					for($i=0;$i<sizeof($application_id);$i++) {
+						$this->branch_manager_model->ConvertApplicationToStudent($application_id[$i]);
+					}
+			}
+			}
+	 redirect('branch_manager_controller/viewApplication');
+		
 	}
 
 
@@ -271,4 +284,151 @@ function Application_by_grade()
 	echo $output;
 
 }
+
+
+// Student Details Function 
+
+public function viewStudentDetails(){
+	//echo "<h1>hello dear </h1>";s
+	$app_id=$this->input->post('id');
+	$details=$this->admin_model->studentMail($app_id);
+	$output="";
+	foreach ($details as $row) {
+		$output.='<div class="block-fluid">
+			
+			<table cellpadding="0" cellspacing="0" width="100%" class="sTable">
+			<tbody>
+			 
+			
+			<tr>
+			<td>';
+			$output.='<img src="'.  base_url().$row['url'].'" class="img-circle" height="100px"  width="100px"></td>';
+			$output.='<td>'. "<span>Name: ".$row['fname']." ".$row['lname'].'</span><br>'.
+							"<span>Gender: ".$row['gender'].'</span><br>'.
+							"<span>D.O.B: ".$row['dob'].'</span><br>'.
+
+							"<span>Language :".$row['language'].'</span><br>'.
+			'</td>
+
+			<td>
+
+					'. "<span>Grade: ".$row['seeking_grade'].'</span><br>'.
+							"<span>Academic Year: ".$row['academic_year'].'</span><br>'.
+							"<span>Religion : ".$row['religion'].'</span><br>'.
+
+							"<span>Blood Group :".$row['blood_group'].'</span><br>'.
+			'
+			
+			</td>
+			</tr>
+				<tr>
+			<td><h5> Parent\'s Details</h5></td>
+				<td>'. "<span>Father Name: ".$row['father_name'].'</span><br>'.
+							"<span>Qualification :".$row['father_qualification'].'</span><br>'.
+							"<span>Occupation :".$row['father_occupation'].'</span><br>'.
+
+							"<span>Mobile:".$row['parents_mobile_number'].'</span><br>'.
+			'</td>
+
+			<td>
+
+					'. "<span>Mother Name: ".$row['mother_name'].'</span><br>'.
+							"<span>Qualification: ".$row['mother_qualification'].'</span><br>'.
+							"<span>Occupation: ".$row['mother_occupation'].'</span><br>'.
+
+							"<span>Email :".$row['parents_email'].'</span><br>'.
+			'
+			
+			</td>
+			</tr>
+
+			<tr>
+			<td><h5> Address</h5></td>
+				<td>'. "<b>Temporary Address ".'</b><br>'.
+							"<span>".$row['temp_address'].'</span><br>'.
+			'</td>
+
+			<td>
+
+					'. "<b>Permanent Address".'</b><br>'.
+							"<span>".$row['parmanent_address'].'</span><br>'.
+			'
+			
+			</td>
+			</tr>
+							<tr>
+			<td><h5> Previous School Details</h5></td>
+				<td>'. "<span>School Name: ".$row['previous_school'].'</span><br>'.
+							"<span>Percentage ".$row['percentage']."%".'</span><br>'.
+							"<span>Phone:".$row['previous_school_phone'].'</span><br>'.
+			'</td>
+
+			<td>
+
+					'. "<b>School  Address".'</b><br>'.
+							"<span>".$row['previous_school_address'].'</span><br>'.
+			'
+			
+			</td>
+			</tr>
+
+
+			
+			</tbody>
+			</table>
+			
+			</div>';
+
+
+	}
+	echo $output;
+
+}
+
+
+public function deleteStudent()
+	{
+		//echo "delete";
+		$application_id=$this->input->post("id");
+		$details=$this->admin_model->studentMail($application_id);
+		$status=$this->branch_manager_model->deleteStudent($application_id);
+		if($status=="true")
+		{
+			foreach ($details as $row) {
+				# code...]
+				$urls=$row('url');
+				unlink($urls);
+			}
+		}
+		echo "hello";
+	}
+
+
+
+	function Student_search()
+{
+	$application =$this->branch_manager_model->Student_search();
+	$output="";
+	foreach ($application as $row) {
+		# code...
+
+		$output.='<tr class="'."s".$row['student_id'].'">
+                                        <td><input type="checkbox" name="order[]" value="'.$row['student_id'].'"/></td>
+                                       <td><a href="#bModal" data-toggle="modal" id="'.$row['student_id'].'"  onClick="viewStudentDetails(this.id)">'.$row["fname"]." ". $row["lname"].'</a></td>
+										<td>'.$row["father_name"].'</a></td>
+										<td>'.$row["seeking_grade"].'</td>
+										<td>'.$row["email"].'</span></td>
+										<td>'.$row["parmanent_address"].'</td>
+										<td class="TAC">
+                                            <a class="action1 tip" title="View Detail" id="'.$row['student_id'].'"  onClick="viewStudentDetails(this.id)" href="#bModal" data-toggle="modal"><span class="icon-ok"></span></a> 
+                                            <a class="action2" id="'.$row['student_id'].'"  href="'.site_url('branch_manager_controller/editStudent'."/".$row['student_id']).'"><span class="icon-pencil"></span></a> 
+                                            <a class="action3 " id="'.$row['student_id'].'" onClick="deletesStudent(this.id)" href="#"><span class="icon-trash"></span></a>
+                                        </td>
+									</tr>';
+	}
+	echo $output;
+}
+
+
+
 }
